Add ACL plugin.
Now it's possible to assign a consumer to one or multiple groups,
configuring the acl plugin to an api allows to whitelist or
blacklist requests from consumers that belong to those groups.

After acl plugin is configured for an api, it becomes mandatory to
provide a consumer for each request, otherwise the request is
forbidden.